melds levelup/leveldown commits into one commit and updates their dependencies appropriately.
