+++ types/level-sublevel/tsconfig.json
@@ -15,8 +15,11 @@
         "types": [],
         "paths": {
             "levelup": [
-                "levelup/v1",
+                "levelup/v1"
             ],
+            "leveldown": [
+                "leveldown/v1"
+            ]
         },
         "noEmit": true,
         "forceConsistentCasingInFileNames": true
