+++ types/leveldown/tslint.json
@@ -1,1 +1,3 @@
-{ "extends": "dtslint/dt.json" }
+{
+  "extends": "dtslint/dt.json"
+}
