+++ types/leveldown/v1/tsconfig.json
@@ -6,7 +6,7 @@
         ],
         "noImplicitAny": true,
         "noImplicitThis": true,
-        "strictNullChecks": false,
+        "strictNullChecks": true,
         "strictFunctionTypes": true,
         "baseUrl": "../../",
         "typeRoots": [
@@ -14,14 +14,11 @@
         ],
         "types": [],
         "paths": {
-            "levelup": [
-                "levelup/v1"
-            ],
-            "levelup/*": [
-                "levelup/v1/*"
-            ],
             "leveldown": [
                 "leveldown/v1"
+            ],
+            "leveldown/*": [
+                "leveldown/v1/*"
             ]
         },
         "noEmit": true,
@@ -29,6 +26,6 @@
     },
     "files": [
         "index.d.ts",
-        "levelup-tests.ts"
+        "leveldown-tests.ts"
     ]
 }
