--- types/leveldown/v4/tsconfig.json
@@ -1,29 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "lib": ["es6"],
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true,
-    "strictFunctionTypes": true,
-    "baseUrl": "../../",
-    "typeRoots": [
-      "../../"
-    ],
-    "types": [],
-    "paths": {
-      "leveldown": [
-        "leveldown/v4"
-      ],
-      "leveldown/*": [
-        "leveldown/v4/*"
-      ]
-    },
-    "noEmit": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "files": [
-		"index.d.ts",
-		"leveldown-tests.ts"
-	]
-}
