--- types/leveldown/v4/tslint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "dtslint/dt.json"
-}
